Add ibmcloud test for unittesting clusterlink APIs

+++ skyflow/network/clusterlink_test.py
@@ -2,13 +2,16 @@ import os
 import sys
 import subprocess
 import shutil
+import argparse
+import logging
+
 projDir = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 sys.path.insert(0,f'{projDir}')
 
 import cluster_linkv2 as clusterlink
 
 from skyflow.cluster_manager import Manager, KubernetesManager
-import logging
+
 logging.basicConfig(
     level=logging.INFO,
     format='%(name)s - %(asctime)s - %(levelname)s - %(message)s')
@@ -23,13 +26,9 @@ if os.path.exists(fabric_cert) != True:
 
 clusterlink_path = os.path.join(clusterlink.CL_DIRECTORY, "clusterlink")
 sys.path.append(clusterlink_path)
-print(sys.path)
 
-from demos.utils.kind import cluster
 from demos.iperf3.kind.iperf3_client_start import testIperf3Client
 
-cluster1 = "peer1"
-cluster2 = "peer2"
 
 cluster1_service_yaml = f"{clusterlink_path}/demos/iperf3/testdata/manifests/iperf3-client/iperf3-client.yaml"
 cluster2_service_yaml = f"{clusterlink_path}/demos/iperf3/testdata/manifests/iperf3-server/iperf3.yaml"
@@ -37,33 +36,79 @@ cluster1_service           = "iperf3-client"
 cluster2_service           = "iperf3-server"
 destPort                   = 5000
 
+# cleanCluster removes all deployments and services 
+def cleanCluster(name : str):
+    subprocess.getoutput(f'kubectl config use-context {name}')
+    subprocess.getoutput('kubectl delete --all deployments')
+    subprocess.getoutput('kubectl delete --all deployments')
+    subprocess.getoutput('kubectl delete --all svc')
+    subprocess.getoutput('kubectl delete --all pods')
+    subprocess.getoutput('kubectl delete --all pvc')
+    subprocess.getoutput('kubectl delete secrets cl-fabric')
+    subprocess.getoutput('kubectl delete secrets cl-peer')
+    subprocess.getoutput('kubectl delete secrets cl-controlplane')
+    subprocess.getoutput('kubectl delete secrets cl-dataplane')
+    subprocess.getoutput('kubectl delete secrets gwctl')
+    subprocess.getoutput('kubectl delete clusterroles cl-controlplane')
+    subprocess.getoutput('kubectl delete clusterrolebindings cl-controlplane')
 
-def cleanup():
+
+def cleanup(cl1, cl2 : str):
     # Cleanup any previous clusters
-    cl1.deleteCluster()
-    cl2.deleteCluster()
-    try :
-        subprocess.check_output(f"rm -rf {clusterlink.CL_DIRECTORY}/kind-{cluster1}", shell=True).decode('utf-8')
-    except:
-        pass
-    try:
-        subprocess.check_output(f"rm -rf {clusterlink.CL_DIRECTORY}/kind-{cluster2}", shell=True).decode('utf-8')
-    except:
-        pass
+    cleanCluster(cl1)
+    cleanCluster(cl2)
+
+    subprocess.check_output(f"rm -rf {clusterlink.CL_DIRECTORY}/{cl1}", shell=True).decode('utf-8')
+    subprocess.check_output(f"rm -rf {clusterlink.CL_DIRECTORY}/{cl2}", shell=True).decode('utf-8')
 
 if __name__ == '__main__':
+    parser = argparse.ArgumentParser(description='Description of your program')
+    parser.add_argument('-e','--env', help='Script command: kind/ibm', required=False, default="test")
+    parser.add_argument('-c','--command', help='Script command: create/delete/none', required=False, default="test")
+    args = vars(parser.parse_args())
+    env = args["env"]
+    command = args["command"]
 
-    cl1 = cluster(cluster1)
-    cl2 = cluster(cluster2)
+    cl1 = None
+    cl2 = None
+    if env == "kind":
+        print("Choosing KIND env")
+        from demos.utils.kind import cluster
+        cl1 = cluster(name="peer1")
+        cl2 = cluster(name="peer2")
+    else:
+        print("Choosing Cloud env")
+        from demos.utils.cloud import cluster
+        cl1 = cluster(name="peer1", zone = "dal10", platform = "ibm")
+        cl2 = cluster(name="peer2", zone = "dal10", platform = "ibm")
 
-    cleanup()
-    # Create a new cluster
-    cl1.createCluster(runBg=True)        
-    cl2.createCluster(runBg=False)  
 
+    if command == "delete":
+        cl1.deleteCluster()
+        cl2.deleteCluster()
+        sys.exit(0)
+    elif command == "create":
+        # Create a new cluster
+        cl1.createCluster(runBg=True)        
+        cl2.createCluster(runBg=False)  
 
-    cluster1_manager= KubernetesManager("kind-"+cluster1)
-    cluster2_manager= KubernetesManager("kind-"+cluster2)
+        cl1.checkClusterIsReady()
+        cl2.checkClusterIsReady()
+        sys.exit(0)
+
+    cl1Name = ""
+    cl2Name = ""
+    if env == "kind":
+        cl1Name = "kind-"+cl1.name
+        cl2Name = "kind-"+cl2.name
+    else:
+        cl1Name = cl1.name
+        cl2Name = cl2.name
+
+    cleanup(cl1Name, cl2Name)
+
+    cluster1_manager= KubernetesManager(cl1Name)
+    cluster2_manager= KubernetesManager(cl2Name)
 
     print("Lauching Clusterlink network on peer1!\n")
     try:
@@ -85,5 +130,7 @@ if __name__ == '__main__':
     clusterlink.export_service(cluster2_service, cluster2_manager, [destPort])
     clusterlink.import_service(cluster2_service, cluster1_manager, cluster2_manager.cluster_name, [destPort])
 
-    testIperf3Client(cl1, cluster1_service, cluster2_service+"-kind-"+cluster2, destPort)
+
+    testIperf3Client(cl1, cluster1_service, cluster2_service+"-"+cl2Name, destPort)
+
